% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repfold.R
\name{addfolder}
\alias{addfolder}
\title{addfolder}
\usage{
addfolder(foldername, uid, M)
}
\arguments{
\item{foldername}{is the name of the folder to be created}

\item{uid}{is the user id of the logged in user}

\item{M}{is the configuration/metadata structure}
}
\description{
Create a new folder to save reports in
}
\details{
Creates a new folder for the user
}
