% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menulib.R
\name{chldtree}
\alias{chldtree}
\title{chldtree}
\usage{
chldtree(cfg, tree)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{tree}{contains the list of ids which are the parent ids that have children}
}
\description{
Get all children of a tree
}
\details{
Returns ids of all the children of requested tree
}
