% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{coltypes}
\alias{coltypes}
\title{coltypes}
\usage{
coltypes(dbconn, tab)
}
\arguments{
\item{dbconn}{is the database handle}

\item{tab}{is the table for which columns are required}
}
\description{
Get column types of a table in a 'MySQL' database
}
\details{
Returns the columns and associated types of the table
}
