% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{datecols}
\alias{datecols}
\title{datecols}
\usage{
datecols(dbconn, tab)
}
\arguments{
\item{dbconn}{is the database handle}

\item{tab}{is the table for which columns are required}
}
\description{
Get the date columns in a table in a 'MySQL' database
}
\details{
Returns the date columns of the table
}
