% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dofilter}
\alias{dofilter}
\title{dofilter}
\usage{
dofilter(tbl, colcond, colval)
}
\arguments{
\item{tbl}{is the name of the tibble}

\item{colcond}{is the column being filtered}

\item{colval}{is the filter value}
}
\description{
Select columns of a 'dplyr' tbl table based a filter
}
\details{
Returns the selected filtered columns of the table
}
