% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dojoin}
\alias{dojoin}
\title{dojoin}
\usage{
dojoin(fact, dim, tj1, tj2)
}
\arguments{
\item{fact}{is the fact table of join}

\item{dim}{is the dimension table of join}

\item{tj1}{is the column of fact table for join}

\item{tj2}{is the column of dimension table for join}
}
\description{
Join two tables
}
\details{
Returns the joined table
}
