% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dosum}
\alias{dosum}
\title{dosum}
\usage{
dosum(tbl, col)
}
\arguments{
\item{tbl}{is the name of the tibble}

\item{col}{are the columns to be aggregated}
}
\description{
Get aggregate value by sum
}
\details{
Returns the aggregated summarised table
}
