% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repfold.R
\name{foldername}
\alias{foldername}
\title{foldername}
\usage{
foldername(cfg, folderid)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{folderid}{is the folder id for which folder name is being asked}
}
\description{
Get report folder name from folder id
}
\details{
Returns report name
}
