% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menulib.R
\name{getchld}
\alias{getchld}
\title{getchld}
\usage{
getchld(mt, pname)
}
\arguments{
\item{mt}{is menu tree which contains all the ids as a tree}

\item{pname}{is the parent name for which child is being looked for}
}
\description{
Get children of a parent (by name)
}
\details{
Returns child menu tree structure
}
