% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menulib.R
\name{getchldbyid}
\alias{getchldbyid}
\title{getchldbyid}
\usage{
getchldbyid(mt, pid)
}
\arguments{
\item{mt}{is menu tree which contains all the ids as a tree}

\item{pid}{is the parent id for which child is being looked for}
}
\description{
Get children of a parent (by id)
}
\details{
Returns child menu tree structure
}
