% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{getcols}
\alias{getcols}
\title{getcols}
\usage{
getcols(dbconn, tab)
}
\arguments{
\item{dbconn}{is the database handle}

\item{tab}{is the table for which columns are required}
}
\description{
Get columns of a table in a 'MySQL' database
}
\details{
Returns the columns of the table
}
