% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initmenu.R
\name{getdimensions}
\alias{getdimensions}
\title{getdimensions}
\usage{
getdimensions(cfg, dims)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{dims}{is the list of dimensions group ids for which details are required}
}
\description{
Get details of dimensions from a list of measure group ids
}
\details{
Returns details of all the dimensions in all the measure groups requested as a dataframe
}
