% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initmenu.R
\name{getdims}
\alias{getdims}
\title{getdims}
\usage{
getdims(cfg, dims)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{dims}{is the list of dimensions ids for which details are required}
}
\description{
Get details of dimensions from a list of ids
}
\details{
Returns details of the dimensions requested as a dataframe
}
