% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initmenu.R
\name{getmeas}
\alias{getmeas}
\title{getmeas}
\usage{
getmeas(cfg, meas)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{meas}{is the list of measure ids for which details are required}
}
\description{
Get details of measures from a list of ids
}
\details{
Returns details of the measures requested as a dataframe
}
