% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initmenu.R
\name{getmeasures}
\alias{getmeasures}
\title{getmeasures}
\usage{
getmeasures(cfg, meas)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{meas}{is the list of measure group ids for which details are required}
}
\description{
Get details of measures from a list of measure group ids
}
\details{
Returns details of all the measures in all the measure groups requested as a dataframe
}
