% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menulib.R
\name{getmnames}
\alias{getmnames}
\title{getmnames}
\usage{
getmnames(m, mid)
}
\arguments{
\item{m}{is the menu details structure (measures / dimensions)}

\item{mid}{is the vector of ids for which names are requested}
}
\description{
Get names from ids
}
\details{
Returns names of ids
}
