% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmodel.R
\name{getmodel}
\alias{getmodel}
\title{getmodel}
\usage{
getmodel(cfg, tabs)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{tabs}{is the list of tables allowed access}
}
\description{
Get the configured data model
}
\details{
Returns table names in the data model, table joins of facts and dimensions, and a matrix of the joins for drawing a chord diagram
}
