% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menulib.R
\name{getparent}
\alias{getparent}
\title{getparent}
\usage{
getparent(cfg, cid)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{cid}{is the child id for whom parent id is required}
}
\description{
Get the parent id of a child
}
\details{
Returns parent id of the child id passed
}
