% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{getprimary}
\alias{getprimary}
\title{getprimary}
\usage{
getprimary(dbconn, tab)
}
\arguments{
\item{dbconn}{is the database handle}

\item{tab}{is the table for which columns are required}
}
\description{
Get the primary key of a table in a 'MySQL' database
}
\details{
Returns the primary key columns of the table
}
