% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gettabs}
\alias{gettabs}
\title{gettabs}
\usage{
gettabs(dbconn)
}
\arguments{
\item{dbconn}{is the database handle}
}
\description{
Get tables in a 'MySQL' database
}
\details{
Returns the set of tables in the database
}
