% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{grpby}
\alias{grpby}
\title{grpby}
\usage{
grpby(tbls, cols)
}
\arguments{
\item{tbls}{is the name of the tibble}

\item{cols}{are the columns to be aggregated}
}
\description{
Aggregate table by a set of columns
}
\details{
Returns the aggregate resulting table
}
