% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{initconfig}
\alias{initconfig}
\title{initconfig}
\usage{
initconfig(configdb)
}
\arguments{
\item{configdb}{is the name of the configuration data base}
}
\description{
Initialise the 'tibbles' of the meta data database
}
\details{
Returns the initialised dplyr configuration handle
}
