% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initmenu.R
\name{initmenu}
\alias{initmenu}
\title{initmenu}
\usage{
initmenu(cfg, mymt)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{mymt}{is a vector of ids which user is allowed access to}
}
\description{
Initialise the session of a user with allowed configuration
}
\details{
Returns a list containing the user's configuration (measures, dimensions)
}
