% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menulib.R
\name{lookupid}
\alias{lookupid}
\title{lookupid}
\usage{
lookupid(M, mnames, grp)
}
\arguments{
\item{M}{is the configuration/metadata structrue}

\item{mnames}{is the list of names for which ids are required}

\item{grp}{is the id group - dimensions or measures}
}
\description{
Look up ids of names
}
\details{
Returns ids corresponding to the names
}
