% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repfold.R
\name{repfold}
\alias{repfold}
\title{repfold}
\usage{
repfold(cfg, uid)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{uid}{is the logged in user id}
}
\description{
Get reports folders of a user id
}
\details{
Returns folder id and names of all folders user has access to
}
