% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repfold.R
\name{replist}
\alias{replist}
\title{replist}
\usage{
replist(cfg, folderid)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{folderid}{is the folder id for which reports list is required}
}
\description{
Get reports list within a folder
}
\details{
Returns report id and title of all reports within the folder id
}
