% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{rowcount}
\alias{rowcount}
\title{rowcount}
\usage{
rowcount(tbl, col)
}
\arguments{
\item{tbl}{is the name of the tibble}

\item{col}{is the column to use for count}
}
\description{
Get row count of a table
}
\details{
Returns the row count of the table
}
