% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{selectcols}
\alias{selectcols}
\title{selectcols}
\usage{
selectcols(tbl, cols)
}
\arguments{
\item{tbl}{is the name of the tibble}

\item{cols}{are the columns to be selected}
}
\description{
Select columns of a 'dplyr' tbl table
}
\details{
Returns the selected columns from the table
}
