% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menulib.R
\name{tabsofids}
\alias{tabsofids}
\title{tabsofids}
\usage{
tabsofids(cfg, ids)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{ids}{is the vector of ids for which table names are required}
}
\description{
Get table names corresponding to ids
}
\details{
Returns the table names of the ids
}
