% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmodel.R
\name{tabstats}
\alias{tabstats}
\title{tabstats}
\usage{
tabstats(cfg, store)
}
\arguments{
\item{cfg}{is the configuration/metadata database connection}

\item{store}{is the data base store}
}
\description{
Get the table statistics for a data base
}
\details{
Returns table row counts, and classification of the tables as small, medium, and large

(Potentially useful for auto idenitification of fact tables)
}
