% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libanlz.R
\name{xgetdims}
\alias{xgetdims}
\title{xgetdims}
\usage{
xgetdims(cfg, metsel)
}
\arguments{
\item{cfg}{is the configuration/metadata structure}

\item{metsel}{is selected measure ids}
}
\description{
Get the dimension groups that can be accessed based on measures selected
}
\details{
Returns dimension groups which can be selected based on measure ids selected
}
