% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libanlz.R
\name{xgetmeas}
\alias{xgetmeas}
\title{xgetmeas}
\usage{
xgetmeas(cfg, dimsel)
}
\arguments{
\item{cfg}{is the configuration/metadata structure}

\item{dimsel}{is selected dimension ids}
}
\description{
Get the measures groups that can be accessed based on dimensions selected
}
\details{
Returns measure groups which can be selected based on dimension ids selected
}
