% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libanlz.R
\name{xmakepicklist}
\alias{xmakepicklist}
\title{xmakepicklist}
\usage{
xmakepicklist(cfg, grp)
}
\arguments{
\item{cfg}{is the configuration/metadata structure}

\item{grp}{is the group of measures or dimensions}
}
\description{
Create a list for pickerInput dropdown
}
\details{
Returns the list with children of the group
}
