% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libanlz.R
\name{xmsel}
\alias{xmsel}
\title{xmsel}
\usage{
xmsel(csel, sel)
}
\arguments{
\item{csel}{is the currently selected ids}

\item{sel}{is the selected/deselected id}
}
\description{
perform a select/deselect operation of an id list
}
\details{
Returns the selected ids after the select/deselect operation
}
